refactor(wire)!: remove resolve_connect_host and move tests to BitcoinSocketAddr
This function was just calling BitcoinSocketAddr, and didn't have any reason to exist.

The tests also tested BitcoinSocketAddr, instead of something inside `conn.rs`, so
I've moved them.
